python/requirements.txt
python-telegram-bot
redis
redis[hiredis]
